% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurites.R
\name{arbour.clusters}
\alias{arbour.clusters}
\alias{arbour.clusters.neuron}
\alias{arbour.clusters.neuronlist}
\alias{arbours}
\alias{arbours.neuron}
\alias{arbours.neuronlist}
\title{Segregate neurite compartments from a neuron/neuronlist based on synapse clustering}
\usage{
arbour.clusters(x, ...)

\method{arbour.clusters}{neuron}(x, ...)

\method{arbour.clusters}{neuronlist}(x, neuronlist = FALSE, ...)

arbours(arbourcluster, ...)

\method{arbours}{neuron}(arbourcluster, ...)

\method{arbours}{neuronlist}(arbourcluster, ...)
}
\arguments{
\item{x}{a neuronlist that has been processed by \code{\link{flow.centrality}}}

\item{...}{additional arguments passed to methods.}

\item{neuronlist}{whether or not to return a neuronlist where each entry is one of the broken up arbours (TRUE)}

\item{arbourcluster}{a neuronlist produced from \code{\link{arbour.clusters}}}
}
\value{
Segmented arbours as a neuronlist object, complete with  synaptic information relevant to that fragment. If a neuronlist is given, subsequently plotting the neuronlist will not reveal the fragments. However, plotting individual neurons double indexed in the lists will. arbours() returns a neuronlist, where each entry is an arbour fragment with a unique skid and name.
}
\description{
Fragment a skeleton into compartments that have been assigned cluster_synapses_within_skeleton() or flow.centrality(). Note that this will break down a neuron regardless of the segregation score calculated across the arbour.
}
\seealso{
\code{\link{neurites}} \code{\link{cluster_synapses_within_skeleton}} \code{\link{seebroken3d}} \code{\link{flow.centrality}} \code{\link{seesplit3d}}
}
