% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape.R
\name{ashape2mesh3d}
\alias{ashape2mesh3d}
\title{Convert an alpha shape to a mesh3D object}
\usage{
ashape2mesh3d(ashape, remove.interior.points = TRUE)
}
\arguments{
\item{ashape}{List of alphashape objects}

\item{remove.interior.points}{whether or not to remove points that are not part of a bounding triangle for the mesh}

\item{...}{additional arguments passed to methods}
}
\value{
A mesh3D object
}
\description{
Convert an alpha shape to a mesh3D object
}
\seealso{
\code{\link{combine.alphashape}} \code{\link{transform3dalphashape}}
}
