% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{assign_cable.polarity}
\alias{assign_cable.polarity}
\title{Assign axon/dendrite split to skeletons}
\usage{
assign_cable.polarity(someneuronlist, brain = nat.flybrains::FCWB, ...)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{brain}{brain template to plot alongside neuron?}

\item{...}{additional arguments passed to methods}
}
\value{
Neuronlist with polarity assignantion marked in the neuron$d dataframe of each neuron object within that neuronlist
}
\description{
manually cycle through and assign a label to points in a neuron to mark out axonic, dednritic and mixed/uncertain cable
}
