% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{assign_lh_neuron}
\alias{assign_lh_neuron}
\title{Assign identitiy to lateral horn neurons}
\usage{
assign_lh_neuron(someneuronlist, most.lhns = NULL, most.lhns.dps = NULL,
  brain = NULL)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{most.lhns}{a dataset containing example neurons of different primary neurite tracts, anatomy groups and cell types. Defaults to the Flycircuit neurons and dye-fills used in Fretcher et al. 2017.}

\item{most.lhns.dps}{a dotprops object of the above. If left NULL, then the function will calculate this, but this is a time consuming operation}

\item{brain}{the brainspace of the someneuronlist. If left NULL, assumes the space is FCWB}

\item{...}{additional arguments passed to methods}
}
\value{
a neuronlist with the best guess for primary neurite tract, anatomy group and cell type listed in its metadata. Quality of this estimation depends on quality of the skeleton objects used in this function and their registration ot FCWB space
}
\description{
Given a neuron, this function assigns it to a lateral horn primary neurite tract, anatomy group and cell type as described by Fretcher et al. 2017.
Note should be taken with what side of the brain the sample group is on. Note that the function will force an assignation even if the neurons in someneuronlist belong to
unidentiified tracts/anatomy group/cell types. Relies on NBlast. Skeletons must have their somas as their root.
}
