% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtubules.R
\name{assign_strahler}
\alias{assign_strahler}
\alias{assign_strahler.neuron}
\alias{assign_strahler.neuronlist}
\title{Assign Strahler stream order to neurites}
\usage{
assign_strahler(x, ...)

\method{assign_strahler}{neuron}(x, ...)

\method{assign_strahler}{neuronlist}(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{...}{Additional arguments passed to nlapply}
}
\description{
Assign Strahler stream order to to neurons / a neuron
}
