% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{assignside}
\alias{assignside}
\title{Assign cell body side based on neuron name}
\usage{
assignside(someneuronlist, ...)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{...}{additional arguments passed to methods}
}
\value{
someneuronlist with cell sidedness in the metadata
}
\description{
Assigns cell body side based in neuron name.
}
