% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{cable.inside.neuropils}
\alias{cable.inside.neuropils}
\alias{cable.inside.neuropils.neuron}
\alias{cable.inside.neuropils.neuronlist}
\alias{points.in.neuropil}
\title{Calculate the cable length inside of different neuropils in a segmented .surf brain}
\usage{
cable.inside.neuropils(x, brain = nat.flybrains::FCWBNP.surf,
  method = c("neurites", "axons", "dendrites"), min.endpoints = 1,
  alpha = 30, ...)

\method{cable.inside.neuropils}{neuron}(x, brain = nat.flybrains::FCWBNP.surf,
  method = c("neurites", "axons", "dendrites"), min.endpoints = 1,
  alpha = 30, ...)

\method{cable.inside.neuropils}{neuronlist}(x,
  brain = nat.flybrains::FCWBNP.surf, method = c("neurites", "axons",
  "dendrites"), min.endpoints = 2, alpha = 30, ...)

\method{points}{in.neuropil}(x, brain, alpha = 30, ...)
}
\arguments{
\item{x}{a set of neurons or, for points.in.neuropil a mxn matrix of 3D points}

\item{brain}{the .surf brainspace in which the neurons are registered, must be segmented into neuropils}

\item{method}{whether to use the neurons' axons or dendrites, or both}

\item{min.endpoints}{the minimum number of endpoints a neuron must have in a neuropil to be counted as included in it}

\item{alpha}{the alpha given to the ashape3d() function to generate neuropil objects by which to calculate point inclusion}

\item{...}{additional arguments passed to methods}
}
\value{
a matrix of 3D points
}
\description{
Calculates the cable length supplied by neurons 'x' to different brain regions defined in 'brain'.
}
