% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_neuropils.R
\name{inside_neuropils}
\alias{inside_neuropils}
\alias{inside_neuropils.neuron}
\title{Calculate the cable length inside of different neuropils in a segmented .surf brain}
\usage{
inside_neuropils(x, brain = nat.flybrains::FCWBNP.surf, method = c("Cable",
  "PRE", "POST"), min.endpoints = 1, alpha = 30, ...)

\method{inside_neuropils}{neuron}(x, brain = nat.flybrains::FCWBNP.surf,
  method = c("Cable", "PRE", "POST"), min.endpoints = 1, alpha = 30, ...)
}
\arguments{
\item{x}{a set of neurons or, for points_in_neuropil a mxn matrix of 3D points}

\item{brain}{the .surf brainspace in which the neurons are registered, must be segmented into neuropils}

\item{method}{whether to calculate cable inside a given neuropil volume, or synapse number (PRE or POST)}

\item{min.endpoints}{the minimum number of endpoints a neuron must have in a neuropil to be counted as included in it}

\item{alpha}{the alpha given to the ashape3d() function to generate neuropil objects by which to calculate point inclusion}

\item{...}{additional arguments passed to methods}
}
\value{
a matrix of 3D points
}
\description{
Calculates the cable length supplied by neurons 'x' to different brain regions defined in 'brain'.
}
