% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid.R
\name{catmaid_annotate_partners}
\alias{catmaid_annotate_partners}
\alias{catmaid_get_annotated_partner}
\alias{reverse.name.side}
\title{Annotate CATMAID partners}
\usage{
catmaid_annotate_partners(partners, pid = 1, conn = NULL, ...)

catmaid_get_annotated_partner(skids, pid = 1, conn = NULL, ...)

reverse.name.side(names)
}
\arguments{
\item{partners}{a vector of two left-right cognates}

\item{pid}{project id (default 1)}

\item{conn}{a catmaid_connection objection returned by catmaid_login. If NULL (the default) a new connection object will be generated using the values of the catmaid.* package options as described in the help for catmaid_login}

\item{...}{additional arguments passed to methods.}

\item{skids}{CATMAID skeleton IDs}

\item{name}{a vector of neuron names}
}
\description{
Intended to use to annotate CATMAID left-right cognates, and fetch them
}
