% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editcatmaid.R
\name{catmaid_connector_nodes}
\alias{catmaid_connector_nodes}
\title{Get information on a CATMAID connector node}
\usage{
catmaid_connector_nodes(connector_id, node = c("presynaptic",
  "postsynaptic"), pid = 1, conn = conn, ...)
}
\arguments{
\item{connector_id}{a data frame of possible mergers between tree nodes for neurons in a CATMAID database, as returned via catmaid_find_likely_merge}

\item{node}{the brain to plot while visualising potential mergers using rgl. Defaults to NULL, no brain plotted.}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch and catmaid::catmaid_get_treenode_detail}
}
\description{
Get information on a CATMAID connector node
}
