% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editcatmaid.R
\name{catmaid_get_server}
\alias{catmaid_get_server}
\title{Get CATMAID server}
\usage{
catmaid_get_server(conn = NULL, ...)
}
\arguments{
\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}
}
\description{
shows the URL for a connection object (see ?catmaid::catmaid_login)
}
