% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editcatmaid.R
\name{catmaid_interactive_name_transfer}
\alias{catmaid_interactive_name_transfer}
\title{Interactively transfer names and annotations between neurons in two different CATMAID instances}
\usage{
catmaid_interactive_name_transfer(
  x = "name:ASB CHECK",
  transfer.annotations = TRUE,
  additional.annotations = "ASB CHECK",
  downsample = 100,
  fafbseg = FALSE,
  duplication.range.nm = 1,
  pid = 1,
  conn = NULL,
  pid2 = 1,
  conn2 = fafb_seg_conn(),
  ...
)
}
\arguments{
\item{x}{a neuron object, or the skeleton ID of a neuron in v14 space}

\item{transfer.annotations}{whether or not to also transfer annotations}

\item{additional.annotations}{an additional annotation to add to the neurons specified by x}

\item{downsample}{if downsample > 1, a sample of number.of.points/downsample is taken from the neuron, and only these points are used to assess duplication. Speeds things up}

\item{fafbseg}{whether or not to use fafbseg::read_brainmaps_meshes on the TODO tag locations and restrict possible duplication to skeletons within the search range and within the cognate auto-segmented volume}

\item{duplication.range.nm}{determines the size of the bounding box around each node in neuron to search for a duplicated. Defaults to 1 nm}

\item{pid}{project id for conn. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{pid2}{project id for conn2. Defaults to 1}

\item{conn2}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
Interactively transfer names and annotations between neurons in two different CATMAID instances
}
