% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editcatmaid.R
\name{catmaid_skeletons_in_bbx}
\alias{catmaid_skeletons_in_bbx}
\alias{catmaid_duplicated}
\title{Search for CATMAID skeletons within a volume}
\usage{
catmaid_skeletons_in_bbx(bbx, min_nodes = 2, pid = 1, conn = NULL,
  ...)

catmaid_duplicated(neuron, tolerance = 0.5, search.range.nm = 1,
  pid = 1, conn = NULL)
}
\arguments{
\item{bbx}{the bounding box (a matrix of 2 rows and 3 columns) describing a search volume}

\item{min_nodes}{the minimum number of nodes a neuron in the search area must have (includes ndoes outside search area)}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch}

\item{neuron}{the neuron object you are thinking about uploading}

\item{tolerance}{how many potentially duplictated nodes  you will tolerate}

\item{search.range.nm}{determines the size of the bounding box around each node in neuron to search for a duplicated. Defaults to 1 nm}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
Programmatically search for skeleton IDs pertaining to neurons within a search volume defined by a bounding box.

Programmatically search for skeleton IDs pertaining to neurons within a search volume defined by a bounding box.
}
