% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{catmaid_upload_neurons}
\alias{catmaid_upload_neurons}
\title{Upload neuron(s) to CATMAID}
\usage{
catmaid_upload_neurons(swc, name = "neuron SWC upload",
  annotations = "SWC upload", pid = 1, conn = NULL,
  max.upload = 10, ...)
}
\arguments{
\item{swc}{a neuron or neuronlist object, or (a) local file path(s) to your saved .swc file(s).}

\item{name}{whatever you want to name your uploaded neurons. If a single character, then it will be added to all uploaded neurons. Else, can be a character vector the same length as swc.}

\item{annotations}{a character vector of annotations, to be added to all ofthe uploaded neurons}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{max.upload}{the maximum number of files that the function will allow you to upload at once}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
Uploads neurons to CATMAID, names them and annotates them.
Please use with caution, as you could be heavily adding to a live tracing environment.
}
