% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catnat-package.R
\docType{package}
\name{catnat-package}
\alias{catnat}
\alias{catnat-package}
\title{Alex Bates' functions for use with nat and rcatmaid}
\description{
Alex Bates' functions for use with nat and rcatmaid
}
\keyword{package}

