% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbysynapse.R
\name{clusterbysynapses}
\alias{clusterbysynapses}
\title{Cluster neurons by pre- and postsynapse positons}
\usage{
clusterbysynapses(someneuronlist, sigma = 1, omega = 1, symmetric = T,
  direction = c(0, 1, 2))
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{sigma}{determines what distances between two synapses are considered close (defaults to 2 um)}

\item{omega}{synapse cluster radius. Defaults to sigma.}

\item{symmetric}{whether to return a symmetric martrix (average of scores between two neurons in both directions)}

\item{...}{additional arguments passed to methods.}
}
\value{
A matrix of similarity scores between inputted neurons, based on synapse positions.
}
\description{
implementation of the algorithm for clustering neurons by synapse location from Schlegel et al. (2016). Assumes neurons are scaled to microns.
}
\details{
From Schneider-Mizell et al. (2016): "We use flow centrality for four purposes. First, to split an arbor into axon and dendrite at the maximum centrifugal SFC, which is a preliminary step for computing the segregation index, for expressing all kinds of connectivity edges (e.g. axo-axonic, dendro-dendritic) in the wiring diagram, or for rendering the arbor in 3d with differently colored regions. Second, to quantitatively estimate the cable distance between the axon terminals and dendritic arbor by measuring the amount of cable with the maximum centrifugal SFC value. Third, to measure the cable length of the main den- dritic shafts using centripetal SFC, which applies only to insect neurons with at least one output syn- apse in their dendritic arbor. And fourth, to weigh the color of each skeleton node in a 3d view, providing a characteristic signature of the arbor that enables subjective evaluation of its identity."
}
\seealso{
\code{\link{plot3d.split}} \code{\link{get.synapses}}
}
