% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterbysynapse.R
\name{clusterbysynapses}
\alias{clusterbysynapses}
\title{Cluster neurons by pre- and postsynapse positions}
\usage{
clusterbysynapses(someneuronlist, sigma = 1, omega = sigma,
  symmetric = T, direction = c(0, 1, 2))
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{sigma}{determines what distances between two synapses are considered close (defaults to 1 um)}

\item{omega}{synapse cluster radius. Defaults to sigma.}

\item{symmetric}{whether to return a symmetric matrix (average of scores between two neurons in both directions)}

\item{direction}{cluster neurons by presynapses (0), postsynapses (1) or both (2)}

\item{...}{additional arguments passed to methods.}
}
\value{
A matrix of similarity scores between inputted neurons, based on synapse positions.
}
\description{
implementation of the algorithm for clustering neurons by synapse location from Schlegel et al. (2016). Assumes neurons are scaled to microns.
}
\seealso{
\code{\link{seesplit3d}} \code{\link{get.synapses}}
}
