% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape.R
\name{combine.alphashape}
\alias{combine.alphashape}
\title{Combine alphashapes generated by the alphashapes package}
\usage{
combine.alphashape(ashapelist, ...)
}
\arguments{
\item{ashapelist}{list of alphashape objects}
}
\value{
An alpha shape object
}
\description{
Combine alphashapes generated by the alphashapes package
}
\references{
Lafarge T, Pateiro-López B, Possolo A, Dunkers J (2014) R Implementation of a Polyhedral Approximation to a 3D Set of Points Using the alpha-Shape. J Stat Softw 56
}
\seealso{
\code{\link{transform3dalphashape}} \code{\link{WriteVTKalphashape}}
}
