% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{connector_URL}
\alias{connector_URL}
\title{Generate URLs to CATMAID connector location}
\usage{
connector_URL(df,
  server = "https://neuropil.janelia.org/tracing/fafb/v14/")
}
\arguments{
\item{df}{data frame that must contain a connector_id column, and x, y, z columns}

\item{server}{CATMAID instance}
}
\description{
Generate URLs to CATMAID connector location
}
