% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerstats.R
\name{node.contribution.to.skeletons}
\alias{node.contribution.to.skeletons}
\alias{synapse.contribution.to.skeletons}
\title{Get data tracers' contributions to skeletons}
\usage{
node.contribution.to.skeletons(x, exclude.authors = NULL, ...)

synapse.contribution.to.skeletons(x, exclude.authors = NULL, direction = NULL)
}
\arguments{
\item{x}{a neuronlist where the name of each neuron is its skeleton ID}

\item{exclude.authors}{a vector of full author names to exclude from result. The result will then print the contributions of non-authors that meet the threshold specified by arguments to suggest_authorship}

\item{...}{argument supplied to suggest_authorship}

\item{direction}{for synapse.contribution.to.skeletons, whether to look at upstream or downstream connections}
}
\description{
Get data tracers' contributions to nodes of skeletons of CATMAID neurons, where the skeletons have been edited e.g. branches removed or added
}
