% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{correctsoma}
\alias{correctsoma}
\title{Re-root neurons to their soma}
\usage{
correctsoma(someneuronlist, brain = NULL, ...)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{brain}{a brain to plot from the nat.templatebrains package e.g. FCWB}

\item{...}{additional arguments passed to methods}
}
\value{
a matrix of 3D points
}
\description{
Cycle through and manually re-root neurons to their soma
}
