% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{create_tracing_googlesheet}
\alias{create_tracing_googlesheet}
\alias{update_tracing_worksheet}
\alias{update_tracing_googlesheet}
\title{Create or update a googlesheet of randomised tracing targets for a given neuron}
\usage{
create_tracing_googlesheet(sheet_title, neuron, skid = neuron$skid,
  polypre = TRUE, axon.dendrite.split = FALSE, random = FALSE)

update_tracing_worksheet(sheet_title, neuron = NULL, skid = neuron$skid,
  ws = "whole neuron input")

update_tracing_googlesheet(sheet_title, neuron = NULL, polypre = FALSE)
}
\arguments{
\item{sheet_title}{title of the googlesheet to be created or updated}

\item{neuron}{a neuron object. If neuron = NULL is given to the update function, data will be updated in the google sheet but any addition or removal of nodes/synapses to the CATMAID neuron will be ignored}

\item{skid}{CATMAID skeleton ID for the neuron}

\item{polypre}{if TRUE, then randomised worksheets for connections laid at the neuron's presynapses are generated}

\item{axon.dendrite.split}{if TRUE and the neuron is labelled in its neuron$d data frame in SWC fashion, then separate worksheets for randomised synapse lists in axon and dendrite are generated}

\item{ws}{name of worksheet in tracing googlesheet to specifically update}
}
\description{
Creates or updates a googlesheet in your drive for a neuron. Worksheets in the googlesheet containin randomised lists of synapses and CATMAID URLs to those synapses,
as well as other meta information to inform tracing. Different worksheets contain separate lists for pre and post synapses, in the whole neuron and in labelled axonic or
dendritic compartments. Randomised pre synapse lists for both presynaptic connector objects for the given neuron and presynaptic connections can be given.
}
