% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtubules.R
\name{distance.from.first.branchpoint}
\alias{distance.from.first.branchpoint}
\alias{distance.from.first.branchpoint.neuron}
\alias{distance.from.first.branchpoint.neuronlist}
\title{Calculate geodesic distance from nodes to a neuron's axon-dendrite
branchpoint}
\usage{
distance.from.first.branchpoint(x, ...)

\method{distance.from.first.branchpoint}{neuron}(x,
  graph.distance = TRUE, ...)

\method{distance.from.first.branchpoint}{neuronlist}(x,
  graph.distance = TRUE, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object that has primary neurites marked (Label =
7) and soma as the root}

\item{...}{Additional arguments passed to nlapply}

\item{graph.distance}{whether to calculate the graph distance (defualt)
between nodes and the primary branchpoint, or the cable length}
}
\description{
alculate geodesic distance from nodes to a neuron's primary,
  axon-dendrite branchpoint
}
