% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{downstream.deletion.test}
\alias{downstream.deletion.test}
\title{Find how many downstream partners tracers have connected up}
\usage{
downstream.deletion.test(someneuronlist, names = c("Alex Bates",
  "Ruairi Roberts"), ...)
}
\arguments{
\item{names}{Full names of tracers to consider}

\item{...}{additional arguments passed to methods.}

\item{someneuron}{a neuronlist object. Should be the downstream partners of (a) neuron(s) of interest}
}
\value{
The percentage of downstream singlet nodes named tracers have contributed to neuron in
}
\description{
Remove points from a neuron, keeping the root node intact
}

