% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{axonic.points}
\alias{axonic.points}
\alias{dendritic.points}
\alias{mixed.points}
\alias{axonic.points.neuron}
\alias{dendritic.points.neuron}
\alias{mixed.points.neuron}
\alias{dendritic.points.neuronlist}
\alias{axonic.points.neuronlist}
\alias{mixed.points.neuronlist}
\alias{axonal.endings}
\alias{dendritic.endings}
\title{Extract axonic/dendritic points from a neuron/neuronlist}
\usage{
axonic.points(someneuronlist, ...)

dendritic.points(someneuronlist, ...)

mixed.points(someneuronlist, ...)

\method{axonic.points}{neuron}(neuron)

\method{dendritic.points}{neuron}(neuron)

\method{mixed.points}{neuron}(neuron)

\method{dendritic.points}{neuronlist}(someneuronlist)

\method{axonic.points}{neuronlist}(someneuronlist)

\method{mixed.points}{neuronlist}(someneuronlist)

axonal.endings(neuron)

dendritic.endings(neuron)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object, which has its axons/dendrites labelled in swc format in its neuron$d dataframes}

\item{...}{additional arguments passed to methods}
}
\value{
a matrix of 3D points
}
\description{
Extract axonic/dendritic points/endpoints from a neuron/neuronlist object
}
