% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{fafb_frags_ids}
\alias{fafb_frags_ids}
\alias{fafb_frags_skeletons}
\alias{fafb_seg_hitlist}
\alias{fafb_seg_tracing_list}
\title{Fetch the up or outgoing auto-traced FAFB fragments from a CATMAID FAFB neuron}
\usage{
fafb_frags_ids(skids, direction = c("incoming", "outgoing"),
  connector_ids = NULL, volume = NULL, ...)

fafb_frags_skeletons(ids, skids = NULL, direction = c("incoming",
  "outgoing"), connector_ids = NULL, read.from = c("CATMAID",
  "Neuroglancer", "local"), ...)

fafb_seg_hitlist(skids, direction = c("incoming", "outgoing"),
  connector_ids = NULL, treat.skids.separately = FALSE, ...)

fafb_seg_tracing_list(skids, direction = c("incoming", "outgoing"),
  connector_ids = NULL, max.nodes = 10, add.links = TRUE,
  unique = TRUE, ...)
}
\arguments{
\item{skids}{neuron skeleton ids}

\item{direction}{whether to fetch putative incoming or outgoing partners}

\item{connector_ids}{restrict your search to only certain connectors. Use if, for example, you want to spatially restrict your search.
Default set to NULL, searches all incoming or outgoing connectors, as specified by direction.}

\item{volume}{volume to which to restrict connector locations}

\item{...}{methods passed to catmaid::catmaid_fetch, catmaid::catmaid_get_connector_table for fafb_frags_ids, and read read methods for fafb_frags_skeletons}

\item{ids}{ids for FAFB segmentations to be read. If skids are given, fafb_frags_ids is called and ids is overlooked.}

\item{treat.skids.separately}{create hitlist, with hits pooled for all skids given (FALSE, default) or per skid given (TRUE)}

\item{max.nodes}{the maximum number of nodes that an extant FAFB partner can have, before we consider using the segmentation for our tracing list.}

\item{add.links}{add links to CATMAID in our tracing list}

\item{unique}{if TRUE, fafb_seg_tracing_list gives each segment only once in the tracing list, the rest of the information is just one example of a putative connection out of the total number, given in the 'hits' column}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}
}
\description{
Fetch skeleton ids or read skeletons from the FAFV Google Segmentation by Peter Li.
This is done by mapping the location of connector nodes in FAFB to the volumetric Google segmentations,
and then their cognate skeletons. Relies on package fafbseg and brainmaps authentication, user list is curated.
}
\details{
fafb_frags_ids returns Neuroglancer IDs for FAFB segments up or downstream of the specified FAFB CATMAID skeleton IDs.
fafb_frags_skeletons reads neurons from Neuroglancer IDs or calls fafb_frags_ids, using either saved skeletons, Neuroglancer brainmaps access or CATMAID access.
fafb_seg_hitlist generates a ranked hitlist of fragments from the given skids, either up or downstream.
fafb_seg_tracing_list goes a bit further and supplies emt information and links to FAFBv14 and the FAFBv14-segmentation instance.
}
