% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{fafb_neuron_details}
\alias{fafb_neuron_details}
\title{Find out what is already known about a neuron's connectivity profile and connected FAFB segments}
\usage{
fafb_neuron_details(skids, direction = c("incoming", "outgoing"),
  connector_ids = connector_ids, volume = NULL, pid = 1,
  conn = NULL, ...)
}
\arguments{
\item{skids}{neuron skeleton ids}

\item{direction}{whether to fetch putative incoming or outgoing partners}

\item{connector_ids}{Optional numeric ids of CATMAID connector nodes that will restrict connections for analysis}

\item{volume}{volume to which to restrict connector locations}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid_get_connector_table}
}
\description{
Set the local path to the location where you have .zip files of FAFB segmented skeletons
}
