% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{fafb_seg_conn}
\alias{fafb_seg_conn}
\title{Log into the v14-seg CATMAID instance for FAFB v14 Adult flybrain segmented skeletonisations}
\usage{
fafb_seg_conn(pid = 1, conn = NULL, ...)
}
\arguments{
\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
Log into the v14-seg CATMAID instance for FAFB v14 Adult flybrain segmented skeletonisations from Peter Li at Google
}
