% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editcatmaid.R
\name{fafbseg_join_connectors_in_brainmaps_volumes}
\alias{fafbseg_join_connectors_in_brainmaps_volumes}
\title{Join free connectors in a CATMAID instance to skeletons if they share a Google Brainmaps volume}
\usage{
fafbseg_join_connectors_in_brainmaps_volumes(neurons,
  putatively.connected.skids, connector.range.nm = 10000, pid = 1,
  conn = NULL, ...)
}
\arguments{
\item{neurons}{a neuronlist object}

\item{putatively.connected.skids}{the skeleton IDs for neurons whose connectors (both incoming and outgoing) may be considered}

\item{connector.range.nm}{the range in nm within which to search for a treenode ID to attach the connector}

\item{pid}{project id. Defaults to 1}

\item{conn}{CATMAID connection object, see ?catmaid::catmaid_login for details}

\item{...}{methods passed to catmaid::catmaid_fetch and catmaid::catmaid_get_treenode_detail}
}
\description{
Join free connectors (the up or downstream node of a conector is joined, where this point belong to a skeleton of 1 node) in a CATMAID instance to skeletons if they share a Google Brainmaps volume
}
