% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{get.skeleton.from.flycircuit}
\alias{get.skeleton.from.flycircuit}
\alias{reroot.flycircuit.neuron}
\alias{Chiang2FCWB}
\title{Fetch Flycircuit neuron skeletons from the Tawian Flycircuit server}
\usage{
get.skeleton.from.flycircuit(fcneurons, ...)

reroot.flycircuit.neuron(neuron)

Chiang2FCWB(x, sex = "F")
}
\arguments{
\item{fcneurons}{a vectors of valid FlyCircuit neuron ids}

\item{...}{additional arguments passed to methods}
}
\value{
A neuronlist of FlyCirucit neurons reigstered in the intersex FCWB brain space
}
\description{
Assigns cell body side based in neuron name.
}
