% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{label_neuron.class}
\alias{label_neuron.class}
\title{See neurons with chosen neuropils}
\usage{
label_neuron.class(x, plotting.brain = nat.flybrains::FCWB,
  regionalised.brain = nat.flybrains::FCWBNP.surf, neuropils = "LH", ...)
}
\arguments{
\item{x}{a neuronlist object, with neurons split into axon and dendrite}

\item{plotting.brain}{brain to plot}

\item{regionalised.brain}{Standard deviation to use in distance dependence of nblast v1 algorithm. Ignored when version=2}

\item{neuropils}{neuropils to plot}

\item{...}{additional arguments passed to methods}
}
\value{
a label_neuron.class
}
\description{
Cycle through neurons, plotting them and chosen neuropils
}
