% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmaking.R
\name{make.anatomical.model}
\alias{make.anatomical.model}
\title{Generate a neuroanatomicla alpha shape from connector and/or tree node data}
\usage{
make.anatomical.model(someneuronlist, substrate = c("connectors", "cable",
  "both"), maxdistance = 10000, groupsize = 100, alpha = 3000,
  auto.selection = T, chosen.points = NULL)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{substrate}{whether to make the model based off of connectors, neuron cable or both}

\item{maxdistance}{for automated cluster identification. Maximum distance at which nodes can be part of a cluster}

\item{groupsize}{an integer number of nearest neighbours to find using nabor::knn()}

\item{alpha}{a single value or vector of values for α, fed to alpshaped3d::ashape3d(). Selection is subsequently interactive}

\item{auto.selection}{whether to try and remove points based on interactively chosen values for 'groupsize' and 'maxdistance'}

\item{chosen.points}{whether to feed the function pre-chosen points. A matrix for 3D points}

\item{selection}{whether or not to interactively select values for maxdistance and groupsize.}

\item{...}{additional arguments passed to methods}
}
\value{
An alphashape object
}
\description{
implementation of the algorithm for clustering neurons by synapse location from Schlegel et al. (2016). Assumes neurons are scaled to microns.
}
