% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmaking.R
\name{make.mirror.transform}
\alias{make.mirror.transform}
\title{Get the transformation matrix from the Morpho::mirror function}
\usage{
make.mirror.transform(x, icpiter = 50, subsample = NULL,
  pcAlign = FALSE, mirroraxis = 1, initPC = TRUE,
  initCenter = TRUE, mc.cores = 2)
}
\arguments{
\item{x}{k x 3 matrix or mesh3d}

\item{icpiter}{Number of iterations to run}

\item{subsample}{integer: use only a subset for icp matching}

\item{pcAlign}{if TRUE, the icp will be preceeded by an alignment of the principal axis (only used if icpiter > 0), currently only works for 3D data}

\item{mirroraxis}{integer: which axis to mirror at}

\item{initPC}{logical: if TRUE the data will be prealigned by its principal axes}

\item{initCenter}{logical: if TRUE and initPC=FALSE, x will be translated to its centroid before mirroring}

\item{mc.cores}{use parallel processing to find best alignment to original shape}

\item{...}{additional arguments passed to methods}
}
\description{
Get the transformation matrix from the Morpho::mirror function
}
