% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{map_fafbsegs_to_neuron}
\alias{map_fafbsegs_to_neuron}
\title{Find out what is already known about a neuron's connectivity profile and connected FAFB segments}
\usage{
map_fafbsegs_to_neuron(
  someneuronlist,
  node.match = 5,
  return.unmatched = FALSE,
  ...
)
}
\arguments{
\item{someneuronlist}{a \code{\link{neuronlist}} or \code{\link{neuron}} object}

\item{node.match}{how many nodes of each neuron in someneuronlist need to be within an auto segmented volume, for it to be said to match.
These nodes all need to be consecutive, in the sense that they must be in the same segment or a branch from that segment. i.e. if a neuron matches with a volume
5 times at diverse points across it arbour, this is thought to be a non-match with a large, proximal auto-traced segment.}

\item{return.unmatched}{defaults to \code{FALSE}. If \code{TRUE}, then a data frame of unmatched Point Numbers for the neuron in question are returned, and their Strahler order.}

\item{...}{Additional arguments passed to \code{fafbseg::\link{brainmaps_xyz2id}} potentiallly including a \code{volume} identifier.}
}
\description{
Set the local path to the location where you have .zip files of FAFB segmented skeletons. This function also very roughly estimates
whether fragment is microtubule containing, what Strahler order it is at, or if it is axonic or dendritic, if the neuron has this marked
}
\section{Volumes}{
 You can choose which auto-segmentation to use by specifying
an optional \code{volume} argument, which will be passed to \code{fafbseg::\link{brainmaps_xyz2id}}. By default the
}

\examples{
\dontrun{
da2=read.neurons.catmaid("glomerulus DA2 right")
fafbseg::with_segmentation("20190805",{
autoseghits=map_fafbsegs_to_neuron(da2)
})
summary(autoseghits)
}
}
