% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{mark.microtubules}
\alias{mark.microtubules}
\alias{prune_microtubules}
\alias{visualise.microtubules}
\title{Functions to assign and visualise microtubule rich and twig portions of a neuron}
\usage{
mark.microtubules(x)

prune_microtubules(x, microtubules = TRUE)

visualise.microtubules(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{microtubules}{whether to return the microtubule containing arbour (TRUE) or twigs (FALSE)}
}
\value{
The neuron/neuronlist object with axon/dendrite info assigned in SWC format to neuron$d
}
\description{
Manually assign the dendrite and axon to neurons / a neuron
}
