% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{most.lhns}
\alias{most.lhns}
\title{A comprehensive lateral horn neuron dataset}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 1226.}
\source{
\url{http://www.flycircuit.tw/}
}
\usage{
most.lhns
}
\description{
A dataset containing all of the FlyCircuit lateral horn neurons that seem to have dendrite in the LH, as well as traced dye-fills from Fretcher et al. 2017
}
\keyword{datasets}
