% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{nblast_bothways}
\alias{nblast_bothways}
\title{NBlast two different sets of neurons forwards and backwards}
\usage{
nblast_bothways(group1, group2 = group1, smat = NULL, sd = 3,
  version = c(2, 1), normalised = FALSE, UseAlpha = FALSE,
  OmitFailures = NA)
}
\arguments{
\item{group1}{a neuronlist object}

\item{group2}{a neuronlist object, defaults to group1}

\item{smat}{the score matrix to be used. See ?nlast.}

\item{sd}{Standard deviation to use in distance dependence of nblast v1 algorithm. Ignored when version=2}

\item{version}{the version of the algorithm to use (the default, 2, is the latest)}

\item{normalised}{whether to divide scores by the self-match score of the query}

\item{UseAlpha}{whether to consider local directions in the similarity calculation (default: FALSE)}

\item{OmitFailures}{Whether to omit neurons for which FUN gives an error. The default value (NA) will result in nblast stopping with an error message the moment there is an eror}

\item{...}{additional arguments passed to methods}
}
\value{
a dotprops object
}
\description{
NBlast two different sets of neurons forwards and backwards, and take the mean of these results
}
