% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neurites.R
\name{neurites}
\alias{neurites}
\alias{neurites.neuron}
\alias{neurites.neuronlist}
\title{Segregate neurite compartments from a neuron/neuronlist}
\usage{
neurites(someneuronlist, fragment = c("axons", "dendrites",
  "primary dendrite", "primary neurite", "nulls"), ...)

\method{neurites}{neuron}(neuron, fragment, ...)

\method{neurites}{neuronlist}(someneuronlist, fragment, ...)
}
\arguments{
\item{someneuronlist}{A neuronlist that has been processed by flow.centrality()}

\item{fragment}{The type of neurite fragment to retrieve. 'Nulls' refers to areas of zero flow. See Schnieder-Mizell et al. (2016)}

\item{...}{additional arguments passed to methods.}
}
\value{
Neurites as a neuronlist object, complete with  synaptic information relevant to that fragment.
}
\description{
Fragment a skeleton into compartments that have been assigned using flow.centrality(). Note that this will break down a neuron regardless of the segregation score calculated by the flow.centrality algorithm.
}
\seealso{
\code{\link{get_connected_skeletons}} \code{\link{skeleton_connectivity_matrix}} \code{\link{flow.centrality}}
}
