% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape.R
\name{neurons.inside}
\alias{neurons.inside}
\title{Find neurites inside of an alphashape}
\usage{
neurons.inside(shape, db, min_nodes = 1)
}
\arguments{
\item{shape}{An alpha shape or mesh3d object}

\item{db}{Neuronlist object that serves as the search database}

\item{min_nodes}{Number of nodes a neuron has to have within the alpha shape
to be selected}

\item{...}{additional arguments passed to methods}
}
\value{
A neuronlist
}
\description{
Find neurites inside of an alphashape
}
\seealso{
\code{\link{combine.alphashape}},
  \code{\link{transform3dalphashape}}, \code{\link[nat]{find.neuron}}
}
