% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_l1.R
\name{open_l1}
\alias{open_l1}
\title{Jump from R into the L1 larval EM data in CATMAID}
\usage{
open_l1(x, s = rgl::select3d(), zoom = 1, open = interactive(),
  scalefac = 1, ...)
}
\arguments{
\item{x}{a 3D shape, neuronlist or neuron object that has been plotted in 3D whose coorinates can be accessed with nat::xyzmatrix()}

\item{s}{selection method. Defaults to rgl selection window.}

\item{zoom}{zoom level to be applied in CATMAID}

\item{open}{whether to open CATMAID window in browser}

\item{scalefac}{the scale factor that has been applied to the neurons/shapes plotted in R}

\item{...}{additional arguments passed to methods.}
}
\value{
Appropriate L1 CATMAID url
}
\description{
Takes you to a region in the Em data chosen by an rgl selection window by the user
}
\details{
CATMAID access required. Data collected and described in cited publication.
}
\references{
Ohyama T, Schneider-Mizell CM, Fetter RD, Aleman JV, Franconville R, Rivera-Alba M, Mensh BD, Branson KM, Simpson JH, Truman JW, et al. (2015) A multilevel multimodal circuit enhances action selection in Drosophila. Nature.
}
