% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{overlap.connectivity.matrix}
\alias{overlap.connectivity.matrix}
\alias{overlap.connectivity.matrix.catmaid}
\title{Generate a connectivity matrix based on euclidean distance between points}
\usage{
overlap.connectivity.matrix(neurons, targets, neuropil = NULL, delta = 1)

overlap.connectivity.matrix.catmaid(neurons, targets, neuropil = NULL,
  delta = 1)
}
\arguments{
\item{neurons}{first set of neurons}

\item{targets}{secodn set of neurons}

\item{...}{additional arguments passed to methods}
}
\value{
a matrix of 3D points
}
\description{
Generates an 'overlap matrix' of overlap scores between neurons in the 'neurons' and 'targets' pools.
For every point in a given neuron in 'neurons', a distance score is calculated to every point in a neuron in 'targets'.
The sum of this score is added to the final output matrix. The score is calculated as e(-d^2/2δ^2), where d is the euclidean distance between the two points,
and δ is the expected distance in um that is considered 'close'.
}
