% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcentrality.R
\name{plot3d.split}
\alias{plot3d.split}
\title{Plot neurons split up by calculating flow centrality}
\usage{
\method{plot3d}{split}(someneuronlist, col = c("blue", "orange", "purple",
  "green", "pink"), WithConnectors = T, WithNodes = F, soma = 100,
  highflow = F)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{col}{colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.}

\item{WithConnectors}{whether ot plot the anatomical location of pre (red) and post (cyan) synapses.}

\item{WithNodes}{whether to plot branch points}

\item{soma}{whether to plot a soma, and what the radius should be}

\item{highflow}{wheather to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)}

\item{primary.dendrite}{Type of object for Deformetrica deformation. See Deformetrica's documentation. Default is appropriate for neuron skeletons.}

\item{...}{additional arguments passed to methods.}
}
\value{
Plots coloured neuron(s)
}
\description{
Plot neurons split up by calculating flow centrality
}
\seealso{
\code{\link{flow.centrality}} \code{\link{get.synapses}}
}

