% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmaking.R
\name{pointsinsidemesh}
\alias{pointsinsidemesh}
\title{Find neurites inside of an mesh3d object}
\usage{
pointsinsidemesh(x, surf, rval = c("logical", "distance", "mesh3d"), ...)
}
\arguments{
\item{x}{3D points}

\item{surf}{A 3D shape, typically a mesh3d object}

\item{rval}{Whetehr to return a logical, distances or a mesh3d object}

\item{...}{additional arguments passed to methods}
}
\value{
A neuronlist
}
\description{
Find neurites inside of an mesh3d object
}
\seealso{
\code{\link{neurons.inside}}
}
