% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{primary.neurite.tracts}
\alias{primary.neurite.tracts}
\title{Lateral horn primary neurite tracts}
\format{An object of class \code{neuronlist} (inherits from \code{list}) of length 30.}
\usage{
primary.neurite.tracts
}
\description{
The known primary neurite tracts that lateral horn neurons can occupy
}
\keyword{datasets}
