% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune.catmaidneuron}
\alias{prune.catmaidneuron}
\alias{prune}
\title{Prune nodes from a catmaid neuron, keeping the synapses}
\usage{
\method{prune}{catmaidneuron}(neuron, target, maxdist, keep = c("near",
  "far"), return.indices = FALSE, ...)
}
\arguments{
\item{neuron}{a neuron object}

\item{target}{Nodes ids for removal}

\item{maxdist}{The threshold distance for keeping points}

\item{keep}{Whether to keep points in x that are near or far from the target}

\item{return.indices}{Whether to return the indices that pass the test rather than the 3D object/points (default FALSE)}

\item{...}{additional arguments passed to methods.}
}
\value{
A pruned neuron object
}
\description{
Prune nodes from a catmaid neuron, keeping the synapses
}
