% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune.in.volume}
\alias{prune.in.volume}
\title{Prune neuron within a mesh3d volume}
\usage{
\method{prune}{in.volume}(x, brain, neuropil = "LH_R", maxdist = 0,
  invert = FALSE)
}
\arguments{
\item{x}{a neuron/neuronlist object that has primary neurites marked (Label =
7) and soma as the root}

\item{brain}{The \code{\link[nat]{hxsurf}} object containing the neuropil of
interest, e.g. \code{\link[nat.flybrains]{FCWBNP.surf}}}

\item{neuropil}{Character vector specifying the neuropil}

\item{maxdist}{The threshold distance for keeping points}

\item{invert}{Logical when \code{TRUE} indicating that points outside the
surface should be pruned.}
}
\description{
Give connector data in a CATMAID neuron the same attributes as
  node data. I.e. adding Label information to indicate compartments such as
  axon and dendrite
}
