% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_by_tag}
\alias{prune_by_tag}
\alias{prune_by_tag.neuron}
\alias{prune_by_tag.catmaidneuron}
\alias{prune_by_tag.neuronlist}
\title{Prune neuron by splitting it at CATMAID tags}
\usage{
prune_by_tag(x, ...)

\method{prune_by_tag}{neuron}(x, tag = "SCHLEGEL_LH", remove.upstream = TRUE, ...)

\method{prune_by_tag}{catmaidneuron}(x, tag = "SCHLEGEL_LH", remove.upstream = TRUE, ...)

\method{prune_by_tag}{neuronlist}(x, tag = "SCHLEGEL_LH", remove.upstream = TRUE, ...)
}
\arguments{
\item{x}{a \code{neuron} or \code{neuronlist} object}

\item{...}{Additional arguments, passed to \code{\link{nlapply}} or eventually to \code{\link{prune_vertices}}}

\item{tag}{a tag that has been assigned in CATMAID}

\item{remove.upstream}{Logical when \code{TRUE} points downstream of the tag(s) are removed, if true, points upstream are removed}
}
\description{
Prune neuron by splitting it at CATMAID tags
}
\details{
Split a neuron based on tags assigned in CATMAID. Remove points either downstream (from the root, must be soma to work properly) of the tagged points(s) or upstream.
}
