% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_in_volume}
\alias{prune_in_volume}
\title{Prune neuron within neuropil volume}
\usage{
prune_in_volume(x, brain, neuropil = "LH_R", maxdist = 0, invert = FALSE)
}
\arguments{
\item{x}{a \code{neuron} object}

\item{brain}{The \code{\link[nat]{hxsurf}} object containing the neuropil of
interest, e.g. \code{\link[nat.flybrains]{FCWBNP.surf}}}

\item{neuropil}{Character vector specifying the neuropil}

\item{invert}{Logical when \code{TRUE} indicating that points outside the
surface should be pruned.}
}
\description{
Prune neuron within neuropil volume
}
\details{
This a slightly odd way of pruning a neuron, \code{x}, first we ask
  which points in x are inside the surface then we keep those parts of x that
  are within \code{maxdist} of those points. \code{maxdist} is therefore
  defined with respect to points in the object, \code{x}, not the surface.
}
