% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_online}
\alias{prune_online}
\alias{prune_online.neuron}
\alias{prune_online.neuronlist}
\title{Prune a neuron interactively}
\usage{
prune_online(x, ...)

\method{prune_online}{neuron}(x, ...)

\method{prune_online}{neuronlist}(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{...}{additional arguments passed to methods}
}
\value{
A pruned neuron/neuronlist object
}
\description{
Remove points from a neuron, keeping the root node intact
}
