% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_strahler.catmaidneuron}
\alias{prune_strahler.catmaidneuron}
\alias{prune_strahler}
\title{Prune a CATMAID neuron by removing segments with a given Strahler order}
\usage{
prune_strahler.catmaidneuron(x, orderstoprune = 1:2, ...)
}
\arguments{
\item{x}{a CATMAID neuron object}

\item{...}{additional arguments passed to methods}
}
\value{
A pruned neuron object
}
\description{
Prune branches by Strahler order from a catmaid neuron, keeping the synapses
}
\seealso{
\code{\link[nat]{prune_strahler}}
}
