% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_vertices.catmaidneuron}
\alias{prune_vertices.catmaidneuron}
\title{Prune vertices from a CATMAID neuron, keeping the synapses}
\usage{
prune_vertices.catmaidneuron(x, verticestoprune, invert = FALSE, ...)
}
\arguments{
\item{x}{a CATMAID neuron object}

\item{...}{additional arguments passed to methods}
}
\value{
A pruned neuron object
}
\description{
Prune nodes from a catmaid neuron, keeping the synapses
}
