% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAFB.R
\name{read.neurons.catmaid.meta}
\alias{read.neurons.catmaid.meta}
\title{Read neurons from CATMAID with extra meta-data}
\usage{
read.neurons.catmaid.meta(skids, meta = c("ItoLee_Lineage",
  "ItoLee_Hemilineage", "Hartenstein_Lineage", "Hartenstein_Hemilineage"),
  sub = " |:", OmitFailures = TRUE, ...)
}
\arguments{
\item{skids}{One or more numeric skeleton ids or a character vector defining
a query (see \code{\link{catmaid_skids}} or examples for the syntax).}

\item{meta}{a vector of meta-annotations to query. The annotations labelled by these meta-annotations will appear as
entries in the \code{data.frame} attached to the returned \code{neuronlist}.}

\item{sub}{what to remove from pulled annotations. Can be set to NULL.}

\item{OmitFailures}{Whether to omit neurons for which \code{FUN} gives an
error. The default value (\code{NA}) will result in \code{\link{nlapply}}
stopping with an error message the moment there is an error. For other
values, see details.}

\item{...}{Additional arguments passed to the \code{\link{catmaid_fetch}}
function}
}
\description{
Use \code{read.neurons.catmaid} to acquire neurons from a CATMAID instance.
In addition, include as meta-data certain annotations that these neurons have. To select which
annotations appear as entries in the neurons' meta-data, select 'meta-annotations' to appear as the
data field. By default, meta-annotations for lineage identities are used, to get the lineage / hemilineage
to which a FAFB Drosophila neuron belongs.
}
