% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{read.neurons.fafbseg}
\alias{read.neurons.fafbseg}
\alias{fafbseg_get_node_count}
\title{Read neurons from the FAFB segmentation instance (temporary)}
\usage{
read.neurons.fafbseg(x, google = FALSE, conn = NULL, ...)

fafbseg_get_node_count(x, read.from = c("CATMAID", "Neuroglancer",
  "local"), ...)
}
\arguments{
\item{x}{typically a skeleton ID in from the CATMAID FAFB v14-seg instance neuroglancer ID.}

\item{google}{if TRUE, we treat x as the name of a Google FAFB segment. If name = FALSE, then x can be skeleton ids (numeric or characters) or the name / annotation for a neuron,
if given as a character starting with 'name:' or 'anotation:' respectively. When Google FAFB fragments are merged, one name is maintained and the other may be found as an annotation. In order to check for this, annotations can also be read.}

\item{conn}{a CATMAID connection object. If NULL catmaid::catmaid::catmaid_login() is called.}

\item{...}{methods passed to catmaid::read.neurons.catmaid}

\item{read.from}{from where to read FAFB segmented skeletons to generate node count data}
}
\description{
Reads neurons (or node count data) from the temporary FAFB segmentation instance.
containing auto-segmented fragments from Peter Li. Neurons can also be read using a neuroglancer server fafbseg::read.neurons.brainmaps
or fafbseg::read_segments2 if from local .zip files
}
