% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid.R
\name{resample.catmaid}
\alias{resample.catmaid}
\alias{resample.catmaid.neuron}
\alias{resample.catmaid.neuronlist}
\title{Resamle a CATMAID neuron}
\usage{
resample.catmaid(someneuronlist, stepsize = 1, ...)

\method{resample}{catmaid.neuron}(neuron, stepsize = 1)

\method{resample}{catmaid.neuronlist}(someneuronlist, stepsize = 1)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object}

\item{stepsize}{The new spacing along the tracing}

\item{...}{additional arguments passed to methods.}

\item{neuron}{A neuron object}
}
\description{
Resample a catmaid neuron so that connector information is retained
}
