% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid.R
\name{resample.catmaidneuron}
\alias{resample.catmaidneuron}
\alias{resample}
\title{Resample a CATMAID neuron}
\usage{
\method{resample}{catmaidneuron}(neuron, stepsize = 1)
}
\arguments{
\item{neuron}{A neuron object}

\item{stepsize}{The new spacing along the tracing}

\item{someneuronlist}{a neuronlist or neuron object}

\item{...}{additional arguments passed to methods.}
}
\description{
Resample a catmaid neuron so that connector information is retained
}
