% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{rescue.dps}
\alias{rescue.dps}
\title{Generate a dps object without dropping small neurons}
\usage{
rescue.dps(someneuronlist, resample = 1, ...)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{resample}{the desired resampling. Smaller neurons will have a lower resampling interval than this.}

\item{...}{additional arguments passed to methods}
}
\value{
a dotprops object
}
\description{
Generates a dotprops object, changing the resample size in order to make sure all neurons are kept
}
