% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustersynapses.R
\name{seebroken3d}
\alias{seebroken3d}
\title{Plot neurons split up by synapse clusters}
\usage{
seebroken3d(neuron, WithConnectors = T, WithNodes = F, soma.size = 100)
}
\arguments{
\item{neuron}{a neuron object that has been modified by flow.centrality}

\item{WithConnectors}{whether ot plot the anatomical location of pre (red) and post (cyan) synapses.}

\item{WithNodes}{whether to plot branch points}

\item{col}{colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.}

\item{soma}{whether to plot a soma, and what the radius should be}

\item{...}{additional arguments passed to methods.}
}
\value{
Plots neuron(s) with arbours coloured by synapse cluster
}
\description{
Plot neurons split up by synapse clusters
}
\seealso{
\code{\link{cluster_synapses_within_skeleton}} \code{\link{seesplit3d}}
}
