% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcentrality.R
\name{seesplit3d}
\alias{seesplit3d}
\alias{splitscan}
\title{Plot neurons split up by flow centrality}
\usage{
seesplit3d(someneuronlist, col = c("blue", "orange", "purple", "green",
  "grey", "pink"), WithConnectors = T, WithNodes = F, soma = 100,
  highflow = F)

splitscan(someneuronlist, col = c("blue", "orange", "purple", "green", "grey",
  "pink"), WithConnectors = T, WithNodes = F, soma = 100, highflow = F,
  Verbose = T, Wait = T, sleep = 0.1, extrafun = NULL,
  selected_file = NULL, selected_col = "black", yaml = TRUE, ...)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object that has been modified by flow.centrality}

\item{col}{colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.}

\item{WithConnectors}{whether ot plot the anatomical location of pre (red) and post (cyan) synapses.}

\item{WithNodes}{whether to plot branch points}

\item{soma}{whether to plot a soma, and what the radius should be}

\item{highflow}{wheather to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)}

\item{...}{additional arguments passed to methods.}

\item{primary.dendrite}{Type of object for Deformetrica deformation. See Deformetrica's documentation. Default is appropriate for neuron skeletons.}
}
\value{
Plots coloured neuron(s)
}
\description{
Plot neurons split up by flow centrality
}
\seealso{
\code{\link{flow.centrality}} \code{\link{get.synapses}}
}
