% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedSampling.R
\name{segments_ranked_pre}
\alias{segments_ranked_pre}
\title{Sampling BLAH}
\usage{
segments_ranked_pre(x, fileout = NULL, omitNAs = FALSE,
  volume = "brainmaps://772153499790:fafb_v14:fafb-ffn1-20190805", ...)
}
\arguments{
\item{x}{a neuron or neuronlist object, or (a) local file path(s) to your saved .swc file(s).}

\item{fileout}{whatever you want to name your uploaded neurons. If a single character, then it will be added to all uploaded neurons. Else, can be a character vector the same length as swc.}

\item{omitNAs}{a character vector of annotations, to be added to all of the uploaded neurons}

\item{...}{methods passed to catmaid::catmaid_fetch}
}
\description{
BLAH
}
