% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{select.points}
\alias{select.points}
\title{Select 3D points in space}
\usage{
\method{select}{points}(points, plot3d = NULL, ...)
}
\arguments{
\item{points}{a mxn matrix}

\item{plot3d}{Object to pass to the rgl::plot3d function}

\item{...}{additional arguments passed to methods}
}
\value{
A mxn matrix
}
\description{
Select and deselect 3D points in space interactively. Black = selected, red = deselected
}
