% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections.R
\name{select_neurons}
\alias{select_neurons}
\alias{deselect_neurons}
\title{Select  neurons in space}
\usage{
select_neurons(someneuronlist, ...)

deselect_neurons(someneuronlist, ...)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{...}{additional arguments passed to methods, i.e. plot3d()}
}
\value{
Selected neurons
}
\description{
Select and deselect neurons in space interactively. Black = selected, red = deselected
}
