% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafbseg.R
\name{set_segmentation_location}
\alias{set_segmentation_location}
\title{Calculate a connectivity similarity score between two connectivity profiles}
\usage{
set_segmentation_location(path)
}
\arguments{
\item{path}{path to FAFB segmentation skeleton .zip files}
}
\description{
Set the local path to the location where you have .zip files of FAFB segmented skeletons
}
