% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftneurons.R
\name{shift.neurons}
\alias{shift.neurons}
\title{Shift neurons in space}
\usage{
shift.neurons(neurons, shift = 1, n = 100)
}
\arguments{
\item{neurons}{a neuron/neuronlist object}

\item{shift}{the size of the shift}

\item{n}{the number of random shifts to perform}
}
\description{
Function for shifting neurons randomly in 3D space. Does not tilt the neuron.
}
