% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shollanalysis.R
\name{sholl_data}
\alias{sholl_data}
\title{Sholl analysis on neuron skeletons}
\usage{
sholl_data(
  neuron,
  start = colMeans(xyzmatrix(neuron)),
  starting.radius = radius.step,
  ending.radius = NULL,
  radius.step = ending.radius/100
)
}
\arguments{
\item{neuron}{a neuron object}

\item{start}{the origin from which spheres are grown for the Sholl analysis}

\item{starting.radius}{the radius of the first sphere. Defaults to the radius step}

\item{ending.radius}{the radius of the last sphere. If NULL the distance to the furthest dendritic point from the start point is taken}

\item{radius.step}{the change in radius between successive spheres. Defaults to one 100th of the radius of the ending sphere}
}
\value{
a data.frame of spheres radii and the number of dendritic intersections at each radius
}
\description{
Functions for Sholl analysis of neuronal skeletons
}
