% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somas.R
\name{soma}
\alias{soma}
\alias{soma.neuron}
\alias{soma.neuronlist}
\alias{somaid}
\alias{somaid.neuron}
\alias{somaid.neuronlist}
\alias{somapos}
\alias{somapos.neuron}
\alias{somapos.neuronlist}
\title{Functions for retrieving soma data}
\usage{
soma(x, ...)

\method{soma}{neuronlist}(x, ...)

\method{soma}{neuron}(x, ...)

somaid(x, ...)

\method{somaid}{neuronlist}(x, ...)

\method{somaid}{neuron}(n)

somapos(x, ...)

\method{somapos}{neuronlist}(x, ...)

\method{somapos}{neuron}(n)
}
\arguments{
\item{...}{additional arguments passed to methods.}
}
\value{
Either the soma's 3D coorindates (soma) its id number (somaid) or its posiiton in the neuron's skeleton (somapos)
}
\description{
Functions for retrieving data amout somata from skeletons, that has been assigned in CATMAID using the tag system
}
\details{
CATMAID access required. Data collected and described in cited publication.
}
\references{
Ohyama T, Schneider-Mizell CM, Fetter RD, Aleman JV, Franconville R, Rivera-Alba M, Mensh BD, Branson KM, Simpson JH, Truman JW, et al. (2015) A multilevel multimodal circuit enhances action selection in Drosophila. Nature.
}

