% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtubules.R
\name{subtree}
\alias{subtree}
\title{Get a subtree in a neuron object as a separate neuron object}
\usage{
subtree(neuron, subtree = 1)
}
\arguments{
\item{neuron}{a neuron object from which to extract a subtree}

\item{subtree}{the number of the subtree to extract}
}
\description{
Get a subtree in a neuron object as a separate neuron object
}
