% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{synapsecolours.neuron}
\alias{synapsecolours.neuron}
\title{Plot a neuron with different coloured synapses depending on partners}
\usage{
synapsecolours.neuron(
  neuron,
  skids = NULL,
  col = "black",
  inputs = T,
  outputs = T,
  printout = F
)
}
\arguments{
\item{neuron}{a neuron to plot}

\item{skids}{the skeleton ids of neurons with which synapses are to be shown. Defaults to all partners.}

\item{col}{the colour of the neuron skeleton}

\item{inputs}{whether or not to show input synapses are to be shown}

\item{outputs}{whether or not output synapses are to be shown}

\item{printout}{whether or not to plot a basic legend to indicate what colours mean which neuron}

\item{...}{additional arguments passed to methods}
}
\description{
Plot a neuron with different coloured synapses depending on partners.
Points indicate inputs synapses, and asterisks indicate output synapses.
}
