% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerstats.R
\name{tracer.plot}
\alias{summarise_contribution}
\alias{tracer.neuron.stats}
\alias{tracer.plot}
\alias{tracer.treenodes.plot}
\alias{write_ack}
\title{Get data and plots describing tracer activity statistics}
\usage{
tracer.plot(names = c("Alex Bates", "Ruairi Roberts", "istvan taisz",
  "Greg Jefferis", "Adam Heath", "Clement Hallou", "Philipp Ranft",
  "Philipp Schlegel"), calc.date = "2013-04-01", from.date = "2016-04-01",
  to.date = Sys.Date(), cumulative = T, ...)

summarise_contribution(skids, auth = 5, ack = 3000, ...)

write_ack(x, ...)

tracer.treenodes.plot(skids = NULL, names = c("Alex Bates",
  "Ruairi Roberts", "istvan taisz", "Greg Jefferis", "Adam Heath",
  "Clement Hallou", "Philipp Ranft", "Philipp Schlegel", "Melissa Ryan"),
  from.date = "2016-04-01", to.date = Sys.Date(), ...)

tracer.neuron.stats(skids, value = c("nodes", "pre", "post"), ...)
}
\arguments{
\item{names}{Full names of tracers}

\item{from.date}{The date from which to plot the graphs}

\item{to.date}{The date to which to plot the graphs}

\item{...}{additional arguments passed to methods.}

\item{skids}{The skids of the neurons to consider}

\item{value}{Whether to look at tree node, pre synapse connector or post synaptic node data.}

\item{clac.date}{The date from which to retrieve data for the graphs}

\item{Cumulative}{Whether to generate a cumulative plot. Defaults to true.}
}
\value{
tracer.plot() returns a plot for cable length traced and connectors placed by named tracers. tracer.treenodes.plot() visualises data on node generation.
}
\description{
Some functions that quickly visualise tracer effort, or retrieve data to do with tracer effort.
}

