% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid.R
\name{update.neuronlist}
\alias{update.neuronlist}
\alias{update.names.neuronlist}
\title{Update a local neuronlist with new CATMAID data}
\usage{
\method{update}{neuronlist}(someneuronlist, skids, ...)

\method{update}{names.neuronlist}(someneuronlist,
  skids = names(someneuronlist), ...)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{skids}{sekeleton IDs}

\item{...}{additional arguments passed to read.neurons.catmaid}
}
\description{
Use to update a large neuronlist quickly by pulling just certain neurons from CATMAID
}
