% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{write.spin.split.swc}
\alias{write.spin.split.swc}
\alias{write.spin.swc}
\title{Write ordered swc}
\usage{
write.spin.split.swc(neuron, file)

write.spin.swc(neuron, file)
}
\arguments{
\item{neuron}{a neuron object}

\item{file}{file path to which to write output}

\item{...}{additional arguments passed to methods}
}
\description{
Write separate, ordered swc files for the MATLAB SPIn package.
One can also write separate files for a neuron's axon, dendrite and other cable
}
