% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape.R
\name{WriteVTKalphashape}
\alias{WriteVTKalphashape}
\title{Write alphashape as a .vtk file}
\usage{
WriteVTKalphashape(
  ashape,
  filename,
  title = filename,
  datatype = c("float", "double")
)
}
\arguments{
\item{ashape}{List of alphashape objects}

\item{filename}{Path to and name of desired file. Should end in .vtk}

\item{title}{Title of the .vtk file}

\item{datatype}{Either float or double}

\item{...}{additional arguments passed to methods}
}
\value{
A .vtk file saved to a given location
}
\description{
Write alphashape as a .vtk file
}
\seealso{
\code{\link{combine.alphashape}} \code{\link{transform3dalphashape}}
}
