% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{assign.cable.polarity}
\alias{assign.cable.polarity}
\title{Assign axon/dendrite split to skeletons}
\usage{
assign.cable.polarity(someneuronlist, ...)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{...}{additional arguments passed to methods}

\item{resample}{the unit lengths to which neurons should be resampled}
}
\value{
Neuronlist with polarity assignantion marked in the neuron$d dataframe of each neuron object within that neuronlist
}
\description{
manually cycle through and assign a label to points in a neuron to mark out axonic, dednritic and mixed/uncertain cable
}
