% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{assign.connector.info}
\alias{assign.connector.info}
\alias{assign.connector.info.neuron}
\alias{assign.connector.info.neuronlist}
\title{Give connector data in a CATMAID neuron the same attributes as node data}
\usage{
assign.connector.info(x, ...)

\method{assign.connector.info}{neuron}(x, ...)

\method{assign.connector.info}{neuronlist}(x, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object that has primary neurites marked (Label = 7) and soma as the root}

\item{...}{Currrently ignored}
}
\description{
Give connector data in a CATMAID neuron the same attributes as
  node data. I.e. adding Label information to indicate compartments such as
  axon and dendrite
}
