% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{assign_cable.polarity}
\alias{assign_cable.polarity}
\title{Assign axon/dendrite split to skeletons}
\usage{
assign_cable.polarity(someneuronlist, brain = nat.flybrains::FCWB, ...)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{brain}{brain template to plot alongside neuron?}

\item{...}{additional arguments passed to methods}
}
\value{
Neuronlist with polarity assignation marked in the \code{neuron$d} \code{dataframe} of each \code{\link{neuron}} object within that \code{\link{neuronlist}}
}
\description{
manually cycle through and assign a label to points in a neuron to mark out axonic, dendritic and mixed/uncertain cable
}
