% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{assign_lh_neuron}
\alias{assign_lh_neuron}
\alias{scan_lh_matches}
\title{Assign identitiy to lateral horn neurons}
\usage{
assign_lh_neuron(someneuronlist, most.lhns = catnat::most.lhns,
  most.lhns.dps = most.lhns.dps,
  most.lhns.pnts.dps = catnat::most.lhns.pnts.dps, brain = NULL)

scan_lh_matches(someassignedneuronlist)
}
\arguments{
\item{someneuronlist}{a neuronlist object}

\item{most.lhns}{a dataset containing example neurons of different primary neurite tracts, anatomy groups and cell types. Defaults to the Flycircuit neurons and dye-fills used in Fretcher et al. 2017, as do the two arguments below}

\item{most.lhns.dps}{a dotprops object of the above}

\item{most.lhns.pnts.dps}{the primary neurite tracts to search against as a dotrpops object}

\item{brain}{the brainspace of the someneuronlist. If left NULL, assumes the space is FCWB}

\item{someassignedneuronlist}{a neuronlist that has been passed through assign_lh_neuron}

\item{...}{additional arguments passed to methods}
}
\value{
a neuronlist with the best guess for primary neurite tract, anatomy group and cell type listed in its metadata. Quality of this estimation depends on quality of the skeleton objects used in this function and their registration ot FCWB space
}
\description{
Given a neuron, this function assigns it to a lateral horn primary neurite tract, anatomy group and cell type as described by Fretcher et al. 2017.
Note should be taken with what side of the brain the sample group is on. Note that the function will force an assignation even if the neurons in someneuronlist belong to
unidentiified tracts/anatomy group/cell types. Relies on NBlast. Skeletons must have their somas as their root. This fucntion provides a good guess, however if it gets a neurons assignment
incorrect, that does not necessarily mean the neuron in question is not covered by the naming system.
}
