% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{cable.inside.neuropils}
\alias{cable.inside.neuropils}
\alias{cable.inside.neuropils.neuron}
\alias{cable.inside.neuropils.neuronlist}
\title{Generate a connectivity matrix based on euclidean distance between points}
\usage{
cable.inside.neuropils(neuron, brain = nat.flybrains::FCWBNP.surf,
  method = c("neurites", "axons", "dendrites"), stepsize = 0.1,
  min.endpoints = 2, alpha = 30, ...)

\method{cable.inside.neuropils}{neuron}(neuron,
  brain = nat.flybrains::FCWBNP.surf, method = c("neurites", "axons",
  "dendrites"), stepsize = 0.1, min.endpoints = 2, alpha = 30)

\method{cable.inside.neuropils}{neuronlist}(someneuronlist,
  brain = nat.flybrains::FCWBNP.surf, method = c("neurites", "axons",
  "dendrites"), stepsize = 0.1, min.endpoints = 2, alpha = 30)
}
\arguments{
\item{...}{additional arguments passed to methods}

\item{neurons}{first set of neurons}

\item{targets}{second set of neurons}
}
\value{
a matrix of 3D points
}
\description{
Generates an 'overlap matrix' of overlap scores between neurons in the 'neurons' and 'targets' pools.
For every point in a given neuron in 'neurons', a distance score is calculated to every point in a neuron in 'targets'.
The sum of this score is added to the final output matrix. The score is calculated as e(-d^2/2δ^2), where d is the euclidean distance between the two points,
and δ is the expected distance in um that is considered 'close'.
}
