% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catnat-package.R
\docType{package}
\name{catnat-package}
\alias{catnat-package}
\alias{_PACKAGE}
\alias{catnat}
\title{catnat: Analyse neuronal morphology and connectivity (extends catmaid and nat)}
\description{
A selection of functions for analysis of neuronal morphology and connectivity created for use alongside the nat and catmaid packages. Many of the functions are driven by the need to analyse the results of annotation of 3D electron microscopy data from the `CATMAID` web application. This package is largely an incubator for work by Alex Bates and is subject to frequent updates.
}
\section{Online Help}{
 See \url{https://jefferislab.github.io/catnat} for
  online documentation.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jefferislab/catnat}
  \item Report bugs at \url{https://github.com/jefferislab/catnat/issues}
}

}
\author{
\strong{Maintainer}: Alex Bates \email{abates@mrc-lmb.cam.ac.uk} (\href{https://orcid.org/0000-0002-1195-0445}{ORCID})

Authors:
\itemize{
  \item Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})
}

}
\keyword{internal}
