% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivity.R
\name{skeleton_connectivity_matrix}
\alias{skeleton_connectivity_matrix}
\alias{connectivity_matrix}
\alias{neuron.heatmap}
\title{Generate connectivity matrix}
\usage{
skeleton_connectivity_matrix(pre, post = NULL, ...)

connectivity_matrix(pre, post = pre, ...)

neuron.heatmap(data, ...)
}
\arguments{
\item{pre}{skeletons from which connections are to be included}

\item{post}{skeletons to which connections from the pre group are to be included. Defaults to the pre group.}

\item{...}{additional arguments passed to methods.}

\item{data}{a connectivity matrix}
}
\value{
A connectivity matrix. The values are synaptic weights, row and column names are neuron names from CATMAID.
}
\description{
Generate a connectivity matrix from neuronlist skeleton data. Can also be used with fragmented neurons, i.e. those generated using the flow.centrality() and neurites() functions.
}
\details{
CATMAID access required.
}
\seealso{
\code{\link{get_connected_skeletons}} \code{\link{neurites}} \code{\link{flow.centrality}}
}
