% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlesheets.R
\name{create_tracing_samplesheet}
\alias{create_tracing_samplesheet}
\alias{update_single_samplesheet}
\alias{update_tracing_samplesheets}
\title{Create or update a googlesheet of randomised tracing targets for a given neuron}
\usage{
create_tracing_samplesheet(neuron, sheet_title = "mystery_neuron",
  folder = "googlesheet", skid = neuron$skid, polypre = TRUE,
  axon.dendrite.split = FALSE, randomise = TRUE)

update_single_samplesheet(sheet_title = "mystery_neuron",
  folder = "googlesheet", neuron = NULL, skid = neuron$skid,
  ws = "whole neuron input")

update_tracing_samplesheets(sheet_title = "mystery_neuron",
  folder = "googlesheet", neuron = NULL, skid = neuron$skid,
  polypre = TRUE)
}
\arguments{
\item{neuron}{a neuron object. If neuron = NULL is given to the update function, data will be updated in the google sheet but any addition or removal of nodes/synapses to the CATMAID neuron will be ignored. If axon.dendrite.split = TRUE, neuron should be passed through catnat::flow.centrality first in order to identify its axon and dendrite}

\item{sheet_title}{title of the googlesheet to be created or updated}

\item{folder}{location to save tracing sheets. Defaults to 'googleshet' which creates a tracign sheet in your ghome folder on googledrive. It can take a while to write a googelsheet. Alternatively, CSVs can be created in a local folder.}

\item{skid}{CATMAID skeleton ID for the neuron}

\item{polypre}{if TRUE, then randomised worksheets for connections laid at the neuron's presynapses are generated}

\item{axon.dendrite.split}{if TRUE and the neuron is labelled in its neuron$d data frame in SWC fashion, then separate worksheets for randomised synapse lists in axon and dendrite are generated}

\item{randomise}{whether to ranomise the synapse sampling list (recommended). If false, the sheet ir organise by  strongest known synaptic parners}

\item{ws}{the individual google worksheet or path to .csv file to update}
}
\description{
Creates or updates a googlesheet in your drive for a neuron. Worksheets in the googlesheet containin randomised lists of synapses and CATMAID URLs to those synapses,
as well as other meta information to inform tracing. Different worksheets contain separate lists for pre and post synapses, in the whole neuron and in labelled axonic or
dendritic compartments. Randomised pre synapse lists for both presynaptic connector objects for the given neuron and presynaptic connections can be given.
}
