% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{get.skeleton.from.flycircuit}
\alias{get.skeleton.from.flycircuit}
\alias{reroot.flycircuit.neuron}
\alias{Chiang2FCWB}
\title{Fetch FlyCircuit neuron skeletons from the Taiwan FlyCircuit server}
\usage{
get.skeleton.from.flycircuit(fcneurons, xform_version = 1, ...)

\method{reroot}{flycircuit.neuron}(x)

Chiang2FCWB(x, female = grepl("-F-", x), xform_version = 1)
}
\arguments{
\item{fcneurons}{a vectors of valid FlyCircuit neuron ids}

\item{xform_version}{the version of nat.templatebrains::xform_brain to use}

\item{...}{additional arguments passed to methods}

\item{x}{some neuron or neuronlist}

\item{female}{whether the neurons to be transformed are female (TRUE) or male (FALSE)}
}
\value{
A neuronlist of FlyCircuit neurons registered in the intersex FCWB brain space
}
\description{
Assigns cell body side based in neuron name.
}
