% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inside_neuropils.R
\name{inside_neuropils}
\alias{inside_neuropils}
\alias{inside_neuropils.neuron}
\alias{inside_neuropils.neuronlist}
\alias{points_in_neuropil}
\alias{in_neuropil}
\alias{in_neuropil.neuron}
\alias{in_neuropil.neuronlist}
\title{Calculate the cable length inside of different neuropils in a segmented .surf brain}
\usage{
inside_neuropils(x, brain, method = c("cable", "PRE", "POST"),
  min.endpoints = 1, alpha = 30, ...)

\method{inside_neuropils}{neuron}(x, brain, method = c("cable", "PRE",
  "POST"), min.endpoints = 1, alpha = 30, ...)

\method{inside_neuropils}{neuronlist}(x, brain, method = c("cable",
  "PRE", "POST"), min.endpoints = 1, alpha = 30, ...)

points_in_neuropil(x, brain, alpha = 30, ...)

in_neuropil(x, brain, neuropil, method = c("cable", "PRE", "POST"),
  min.endpoints = 1, alpha = 30, ...)

\method{in_neuropil}{neuron}(x, brain, neuropil, method = c("cable",
  "PRE", "POST"), min.endpoints = 1, alpha = 30, ...)

\method{in_neuropil}{neuronlist}(x, brain = nat.flybrains::FCWBNP.surf,
  neuropil = "LH_R", method = c("cable", "PRE", "POST"),
  min.endpoints = 1, alpha = 30, ...)
}
\arguments{
\item{x}{a set of neurons or, for points_in_neuropil a mxn matrix of 3D points}

\item{brain}{the .surf brainspace in which the neurons are registered, must be segmented into neuropils}

\item{method}{whether to calculate cable inside a given neuropil volume, or synapse number (PRE or POST)}

\item{min.endpoints}{the minimum number of endpoints a neuron must have in a neuropil to be counted as included in it}

\item{alpha}{the alpha given to the ashape3d() function to generate neuropil objects by which to calculate point inclusion}

\item{...}{additional arguments passed to methods}

\item{neuropil}{the code for the neuropil volume to be searched.}
}
\value{
a matrix of 3D points
}
\description{
Calculates the cable length supplied by neurons 'x' to different brain regions defined in 'brain'.
}
