% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtubules.R
\name{mark.microtubules}
\alias{mark.microtubules}
\alias{mark.microtubules.neuron}
\alias{mark.microtubules.neuronlist}
\alias{prune_microtubules}
\alias{prune_microtubules.neuron}
\alias{prune_microtubules.neuronlist}
\alias{visualise.microtubules}
\alias{microtubules.errors}
\title{Functions to assign and visualise microtubule rich and twig portions of a
neuron}
\usage{
mark.microtubules(x, ...)

\method{mark.microtubules}{neuron}(x)

\method{mark.microtubules}{neuronlist}(x, ...)

prune_microtubules(x, ...)

\method{prune_microtubules}{neuron}(x, microtubules = TRUE)

\method{prune_microtubules}{neuronlist}(x, microtubules = TRUE)

visualise.microtubules(x, soma = TRUE, WithConnectors = FALSE, ...)

microtubules.errors(skid)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{...}{Additional arguments passed to nlapply}

\item{microtubules}{whether to return the microtubule containing arbour
(TRUE) or twigs (FALSE)}

\item{soma}{for \code{visualise.microtubules} whether to show the soma}

\item{WithConnectors}{for \code{visualise.microtubules} whether to show
connectors}

\item{skid}{skeleton ID of CATMAID neuron for checking whether there are
presynapses marked as being on a microtubule-lacking twig}
}
\value{
The neuron/neuronlist object with axon/dendrite info assigned in SWC
  format to neuron$d
}
\description{
Manually assign the dendrite and axon to neurons / a neuron
}
