% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{napplyTransform}
\alias{napplyTransform}
\alias{napplyTransform.neuron}
\alias{napplyTransform.neuronlist}
\title{Apply a transform to a neuron/neuronlist}
\usage{
napplyTransform(x, trafo, inverse = F, ...)

\method{napplyTransform}{neuron}(x, trafo, inverse = F, ...)

\method{napplyTransform}{neuronlist}(x, trafo, inverse = F, ...)
}
\arguments{
\item{x}{a neuron/neuronlist object}

\item{trafo}{A valid transformation for \code{Morpho::\link{applyTransform}}}

\item{inverse}{whether to calculate the inverse of trafo}

\item{...}{additional arguments passed to methods}
}
\value{
Transformed neuron/neuronlist object
}
\description{
Apply a transform to a neuron/neuronlist object using Morpho::applyTransform
}
