% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flycircuit.R
\name{napplyTransform}
\alias{napplyTransform}
\alias{napplyTransform.neuronlist}
\title{Apply a transform to a neruon/neuronlist}
\usage{
napplyTransform(someneuronlist, trafo, inverse = F, ...)

\method{napplyTransform}{neuronlist}(someneuronlist, trafo, inverse = F, ...)
}
\arguments{
\item{someneuronlist}{a neuron/neuronlist object}

\item{trafo}{A valid transformation for Morpho:applyTransform}

\item{inverse}{whether to calculate the inverse of trafo}

\item{...}{additional arguments passed to methods}
}
\value{
Transformed neuron/neuronlist object
}
\description{
Apply a transform to a neruon/neuronlist object using Morpho::applyTransform
}
