% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphashape.R, R/modelmaking.R
\name{neurons.inside}
\alias{neurons.inside}
\alias{pointsinsidemesh}
\title{Find neurites inside of an alphashape}
\usage{
neurons.inside(alpha, db, synapse = "BOTH", degree = NULL)

pointsinsidemesh(x, surf, ..., rval = c("logical", "distance", "mesh3d"))
}
\arguments{
\item{db}{Neuronlist object that serves as the search database}

\item{x}{3D points}

\item{surf}{A 3D shape, typically a mesh3d object}

\item{...}{additional arguments passed to methods}

\item{rval}{Whetehr to return a logical, distances or a mesh3d object}

\item{shape}{An alpha shape or mesh3d object}

\item{min_nodes}{Number of nodes a neuron has to have within the alpha shape to be selected}

\item{...}{additional arguments passed to methods}
}
\value{
A neuronlist

A neuronlist
}
\description{
Find neurites inside of an alphashape

Find neurites inside of an mesh3d object
}
\seealso{
\code{\link{combine.alphashape}} \code{\link{transform3dalphashape}}

\code{\link{neurons.inside}}
}
