% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{primary.neurite}
\alias{primary.neurite}
\alias{primary.neurite.neuron}
\alias{primary.neurite.neuronlist}
\alias{longestpathfromsoma}
\title{Returns the primary neurite of a neuron}
\usage{
primary.neurite(someneuron, k = 100)

primary.neurite.neuron(neuron, resample = 1, ...)

primary.neurite.neuronlist(someneuronlist, ...)

longestpathfromsoma(n, UseStartPoint = TRUE, SpatialWeights = TRUE,
  invert = FALSE, rval = c("neuron", "length", "ids"), model = NULL)
}
\arguments{
\item{neuron}{a neuron object}

\item{resample}{The newspacing with which to evenly resample each neuron. Can be set to F to prevent resampling.}

\item{...}{additional arguments passed to methods}
}
\value{
A neuron pruned to its primary dendrite
}
\description{
Returns the primary neurite of a neuron, defined as the cable between soma and first branch point
}
