% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronlist.R
\name{primary.neurite}
\alias{primary.neurite}
\alias{primary.neurite.neuron}
\alias{primary.neurite.neuronlist}
\title{Returns the primary neurite of a neuron}
\usage{
primary.neurite(someneuron, k = 100)

primary.neurite.neuron(neuron, ...)

primary.neurite.neuronlist(someneuronlist, ...)
}
\arguments{
\item{neuron}{a neuron object}

\item{...}{additional arguments passed to methods}
}
\value{
A neuron pruned to its primary dendrite
}
\description{
Returns the primary neurite of a neuron, defined as the cable between soma and first branch point
}

