% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_distal}
\alias{prune_distal}
\title{Prune nodes from a neuron, keeping the root node}
\usage{
prune_distal(neuron, bad_vertex_labels, invert = FALSE, ...)
}
\arguments{
\item{neuron}{a neuron object}

\item{bad_vertex_labels}{Nodes ids for removal}

\item{...}{additional arguments passed to methods.}
}
\value{
A pruned neuron object
}
\description{
Remove points from a neuron, keeping the root node intact
}
