% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_in_volume}
\alias{prune_in_volume}
\alias{prune_in_volume.neuron}
\alias{prune_in_volume.neuronlist}
\title{Prune neuron within neuropil volume}
\usage{
prune_in_volume(x, brain = nat.flybrains::FCWBNP.surf, neuropil = "LH_R",
  invert = TRUE, ...)

\method{prune_in_volume}{neuron}(x, brain = nat.flybrains::FCWBNP.surf,
  neuropil = "LH_R", invert = TRUE, ...)

\method{prune_in_volume}{neuronlist}(x, brain = nat.flybrains::FCWBNP.surf,
  neuropil = "LH_R", invert = TRUE, ...)
}
\arguments{
\item{x}{a \code{neuron} object}

\item{brain}{The \code{\link[nat]{hxsurf}} object containing the neuropil of
interest, e.g. \code{\link[nat.flybrains]{FCWBNP.surf}}}

\item{neuropil}{Character vector specifying the neuropil}

\item{invert}{Logical when \code{TRUE} indicating that points outside the}

\item{...}{Additional arguments for methods (eventually passed to prune.default)
surface should be pruned.}
}
\description{
Prune neuron within neuropil volume
}
\details{
Prune neuron inside subvolume of a segmented brain object.
}
