% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmaid.R
\name{resample.catmaidneuron}
\alias{resample.catmaidneuron}
\alias{resample}
\title{Resample a CATMAID neuron}
\usage{
resample.catmaidneuron(x, stepsize = 1, ...)
}
\arguments{
\item{x}{A neuron or neuronlist object}

\item{stepsize}{The new spacing along the tracing}

\item{...}{additional arguments passed to methods.}
}
\description{
Resample a catmaid neuron so that connector information is
  retained
}
\details{
\bold{NB} the connector \code{treenode_id} column will retain all
  the original CATMAID node ids and will therefore no longer match the
  PointNo identifiers for segments in the resampled neuron.
}
