% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcentrality.R
\name{seesplit3d}
\alias{seesplit3d}
\alias{splitscan}
\title{Plot neurons split up by flow centrality}
\usage{
seesplit3d(someneuronlist, col = c("blue", "orange", "purple", "green",
  "grey", "pink"), WithConnectors = T, WithNodes = F, soma = 100,
  highflow = F)

splitscan(someneuronlist, col = c("blue", "orange", "purple", "green", "grey",
  "pink"), WithConnectors = T, WithNodes = F, soma = 100, highflow = F,
  Verbose = T, Wait = T, sleep = 0.1, extrafun = NULL,
  selected_file = NULL, selected_col = "black", yaml = TRUE, ...)
}
\arguments{
\item{someneuronlist}{a neuronlist or neuron object that has been modified by flow.centrality}

\item{col}{colours of sections. Defaults to orange or axons, green for primary dendrite, blue for dendrites and pink for nodes with no flow.}

\item{WithConnectors}{whether ot plot the anatomical location of pre (red) and post (cyan) synapses.}

\item{WithNodes}{whether to plot branch points}

\item{soma}{whether to plot a soma, and what the radius should be}

\item{highflow}{wheather to plot the nodes of highest (with in one standard deviation less than maximum) flow centrality (pink points)}

\item{Verbose}{logical indicating that info about each selected neuron should be printed (default TRUE)}

\item{Wait}{logical indicating that there should be a pause between each displayed neuron}

\item{sleep}{time to pause between each displayed neuron when Wait=TRUE}

\item{extrafun}{an optional function called when each neuron is plotted, with two arguments: the current neuron name and the current selected neurons}

\item{selected_file}{an optional path to a yaml file that already contains a selection}

\item{selected_col}{the color in which selected neurons (such as those specified in selected_file) should be plotted}

\item{yaml}{a logical indicating that selections should be saved to disk in (human-readable) yaml rather than (machine-readable) rda format}

\item{...}{additional arguments passed to methods.}
}
\value{
Plots coloured neuron(s)
}
\description{
Plot neurons split up by flow centrality
}
\seealso{
\code{\link{flow.centrality}} \code{\link{get.synapses}}
}
