% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankedSampling.R
\name{segments_ranked_pre}
\alias{segments_ranked_pre}
\alias{segments_ranked_post}
\title{Ranked sampling}
\usage{
segments_ranked_pre(
  x,
  fileout = NULL,
  omitNAs = FALSE,
  volume = "brainmaps://772153499790:fafb_v14:fafb-ffn1-20190805",
  ...
)

segments_ranked_post(
  x,
  fileout = NULL,
  omitNAs = FALSE,
  volume = "brainmaps://772153499790:fafb_v14:fafb-ffn1-20190805",
  ...
)
}
\arguments{
\item{x}{neuron or neuronlist. given by a CATMAID skid, skid list, or annotation. Synaptic partners of these neurons will be returned.}

\item{fileout}{character. defines the file to write out the generated spreadsheet in csv format}

\item{omitNAs}{logical. Should NAs be omitted? FALSE by default}

\item{...}{catmaid::catmaid_get_connector_table for segments_ranked_pre, catmaid::catmaid_get_connectors_between for segments_ranked_post}
}
\value{
A \code{list}, most importantly the columns: "segment"  - google segment ID, "n" - number of overlaps with query neuron connections, "FAFB.link" and "FAFBseg.link" URLs to position of the connection in the regular and the Google seg CATMAID environments, respectively.
}
\description{
Generating a ranked sampling spreadsheet based on number of overlapping synapses with segments for CATMAID neurons or neuron lists. The function relies on the brainmaps API from Google. The result depends on marked up pre- or postsynapses.
}
\examples{
\donttest{
DA2_up_segs = segments_ranked_pre("glomerulus DA2 right")
View(DA2_up_segs)
segments_ranked_post(38885, fileout = "single_DA2_downstream_segs.csv", omitNAs = TRUE)
}
}
