% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracerstats.R
\name{tracer.plot}
\alias{tracer.plot}
\alias{tracer.treenodes.plot}
\alias{tracer.neuron.stats}
\title{Get data and plots describing tracer activity statistics}
\usage{
tracer.plot(names = c("Alex Bates", "Ruairi Roberts", "Greg Jefferis",
  "Clement Hallou", "Philipp Ranft", "Philipp Schlegel", "Fiona Love",
  "Amelia Edmondson-Stait"), calc.date = "2016-04-01",
  from.date = "2016-04-01", to.date = Sys.Date(), cumulative = T,
  ...)

tracer.treenodes.plot(skids = NULL, names = c("Alex Bates",
  "Ruairi Roberts", "Greg Jefferis", "Clement Hallou", "Philipp Ranft",
  "Philipp Schlegel", "Fiona Love", "Amelia Edmondson-Stait"),
  from.date = "2016-04-01", to.date = Sys.Date(), ...)

tracer.neuron.stats(skids, value = c("nodes", "pre", "post"), ...)
}
\arguments{
\item{names}{Full names of tracers}

\item{calc.date}{The date from which to retrieve data for the graphs}

\item{from.date}{The date from which to plot the graphs}

\item{to.date}{The date to which to plot the graphs}

\item{cumulative}{Whether to generate a cumulative plot. Defaults to true.}

\item{...}{additional arguments passed to methods.}

\item{skids}{The skids of the neurons to consider}

\item{value}{Whether to look at tree node, pre synapse connector or post synaptic node data.}
}
\value{
tracer.plot() returns a plot for cable length traced and connectors placed by named tracers. tracer.treenodes.plot() visualises data on node generation.
}
\description{
Some functions that quickly visualise tracer effort, or retrieve data to do with tracer effort.
}
